requirements.txt
pandas
python-dotenv
pyodbc
requests
psycopg2
atproto